Replaced license with GNU GPLv3
I decided to switch to the GNU GPLv3 from the MIT License. I find that
the GPL will better protect this software. The main difference between
the two that I care about is that any derivative works must also stay
open-source. As the creator of Budg, I want to be able to see changes
made by the community and use them to make Budg the best and most
useful tool it can be.