creates plans and budgits
Due to the ubiquitous use of the term "budget", I created two new terms:
plans and budgits. Plans are nested dictionaries that contain the
percentages of the paycheck that belong to each line item. Budgits are
nested dictionaries that contain the dollar amount that is allocated to
each line item.

Functions are redefined:
- budget and parseBudget are removed
- readFile gets a configparser from the ini file
- parsePlan converts configparser to a plan
- calcBudgit takes a check and plan, and returns a budgit
- printBudgit will print the budgit to stdout