Fix test cases
The test cases were bugged. The methods all returned 0, which was always
less than the amount they were given, except for in the case of the
negatives.

This is a great example of why testing is important. It looked like all
of my tests were passing except one, but what was really happening was
all of my tests were bad.