Add mechanics for console interface
The console interface now works. It can be used to budget paychecks
according to your plan. It is ready for more features and commands to be
added.

I also refactored some code. Plans now go straight from the disk to a
dictionary through configparser with a single function call instead of
two. Sessions are now handled outside of main. The totalize function now
takes a list of strings, converts them to floats, and sums them up.