adds Budg class and implementation, cleans smell
wow this is a big commit
1. the Budg session class is created. Main simply spins up a Budg
session and manages the program through the session interface.
2. The Budg session is an object that holds budg configuration and plan
data. It handles all input during execution.
3. I got rid of uint. I thought that uint made sense, but the internet
says to avoid them because of clarity. Since it is not
crucial for the penny count to be unsigned, it makes sense to store it
as a regular int.
4. I explicitly reference std::string instead of typedef std::string
string. Just to make things clear and standard.
5. I changed any upperCamelCase to under_scores.
6. I switched from cstdio to iostream. Again, to make things standard
and easy. Personally, I really enjoy the c libraries, but this is not c.