move source files into budg dir
the src folder was a little unnecessary, and the ini file inside is not
really source. With this naming convention, I sill not need to import
"src" in any other files, which would be rather silly.

I figured a more simplistic project structure would help me get going on
this project.