move plan dir to userhome/Documents
type: enhancement

Keeping plans in .config meant that the user had to make hidden files
and folders visible in order to work on their budget. This was not
ideal.
I moved the directory where plans are kept to Documents in the user's
home directory. Also improved the readFile function by getting rid of
some unneccessary variables and reorganizing the order of execution. It
will now make sure the plan file exists before initiating the
ConfigParser. Added better documentation with typehints and docstring.

re: #25 Move plans to Documents