update pipeline
I do not think black will actually change anything in the commit, so it
should be run as a checker.

Also, I just learned that flake8 and black are not *entirely* compatible
because of E203 (an issue I actually discovered today at work and didn't
even realize until now).

Also, since these are just checking code, let's save some build minutes
by running in parallel (idk if that's how it works but why not).