begin migrating to golang
I believe that golang is much better suited to this tool, as it avoids
all of the wonky install pieces of using python. You no longer need to
run this script through an interpreter. You can build an executable and
install it just like you would any other piece of software. You would
also get the performance benefit of using a compiled language.